requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
